requirements.txt
Django==5.0.2
python-decouple==3.8
mysqlclient==2.2.4
djangorestframework==3.14
djangorestframework-simplejwt==5.3.1
django-rest-passwordreset==1.4.0